+++ includes/class-tmsm-woocommerce-booking-thalasso.php
@@ -103,6 +103,7 @@ class Tmsm_Woocommerce_Booking_Thalasso {
 		 * Elementor overrides
 		 */
 		if ( function_exists( '_is_elementor_installed' ) ) {
+			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-noaccommodationpackageprice.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-accommodationpackageprice.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-packageprice.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-bookresaweburl.php';